Moved vtk files
The Writevtk class was in willard_chandler.py
